% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_XPASS.R
\name{predict_XPASS}
\alias{predict_XPASS}
\title{Construct PRS from the XPASS output.}
\usage{
predict_XPASS(pm, file_predgeno)
}
\arguments{
\item{pm}{The output `mu` obtained from the `XPASS` function, which stores the posterior means computed by XPASS.}

\item{file_predgeno}{Prefix of the prediction samples' genotypes in plink format.}
}
\value{
A data frame storing the PRS generated by using mu1, mu2, mu_XPASS1, and mu_XPASS2, respectively.
}
\description{
This function use the XPASS posterior means to construct PRS for the prediction samples.
}
\examples{
See vignette.
}
